Move Functionality in Cassandra
