+++ cassandra/src/main/scala/geotrellis/spark/io/cassandra/CassandraLayerReader.scala
@@ -30,6 +30,7 @@ class CassandraLayerReader(val attributeStore: AttributeStore, instance: Cassand
   extends FilteringLayerReader[LayerId] {
 
   val defaultNumPartitions = sc.defaultParallelism
+  def sparkContext: SparkContext = sc
 
   def read[
   K: AvroRecordCodec: Boundable: JsonFormat: ClassTag,
